feat: automatically name training dir
