+++ mlpf/utils.py
@@ -1,0 +1,16 @@
+from pathlib import Path
+import datetime
+import platform
+
+
+def create_experiment_dir(prefix=None, suffix=None):
+    if prefix is None:
+        train_dir = Path("experiments") / datetime.datetime.now().strftime("%Y%m%d_%H%M%S_%f")
+    else:
+        train_dir = Path("experiments") / (prefix + datetime.datetime.now().strftime("%Y%m%d_%H%M%S_%f"))
+
+    if suffix is not None:
+        train_dir = train_dir.with_name(train_dir.name + "." + platform.node())
+
+    train_dir.mkdir(parents=True)
+    return str(train_dir)
